@@ -69,15 +69,15 @@
 		"serialization/serialization": "4.0.0",
 		"spomky-labs/base64url": "2.0.1",
 		"spomky-labs/cbor-php": "1.0.8",
-		"symfony/console": "5.0.8",
-		"symfony/debug": "4.4.8",
-		"symfony/process": "5.0.8",
-		"symfony/polyfill-php73": "1.17.0",
-		"symfony/service-contracts": "2.0.1",
-		"symfony/translation": "5.0.8",
-		"symfony/translation-contracts": "2.0.1",
-		"symfony/validator": "5.0.8",
-		"symfony/var-dumper": "5.0.8",
+		"symfony/console": "5.1.2",
+		"symfony/debug": "4.4.10",
+		"symfony/process": "5.1.2",
+		"symfony/polyfill-php73": "1.17.1",
+		"symfony/service-contracts": "2.1.2",
+		"symfony/translation": "5.1.2",
+		"symfony/translation-contracts": "2.1.2",
+		"symfony/validator": "5.1.2",
+		"symfony/var-dumper": "5.1.2",
 		"web-auth/cose-lib": "3.2.6",
 		"web-auth/metadata-service": "3.2.6",
 		"web-auth/webauthn-lib": "3.2.6",
@@ -123,6 +123,8 @@
 		"paragonie/random_compat": "9.99.99",
 		"symfony/polyfill-ctype": "1.99",
 		"symfony/polyfill-intl-idn": "1.17.0",
+		"symfony/polyfill-intl-grapheme": "1.17.1",
+		"symfony/polyfill-intl-normalizer": "1.17.1",
 		"symfony/polyfill-mbstring": "1.99",
 		"symfony/polyfill-php72": "1.12.0"
 	},
