@@ -16,10 +16,12 @@
         }
     ],
     "require": {
-        "php": "^7.2.5",
+        "php": ">=7.2.5",
         "symfony/polyfill-mbstring": "~1.0",
         "symfony/polyfill-php73": "^1.8",
-        "symfony/service-contracts": "^1.1|^2"
+        "symfony/polyfill-php80": "^1.15",
+        "symfony/service-contracts": "^1.1|^2",
+        "symfony/string": "^5.1"
     },
     "require-dev": {
         "symfony/config": "^4.4|^5.0",
@@ -41,6 +43,7 @@
     },
     "conflict": {
         "symfony/dependency-injection": "<4.4",
+        "symfony/dotenv": "<5.1",
         "symfony/event-dispatcher": "<4.4",
         "symfony/lock": "<4.4",
         "symfony/process": "<4.4"
@@ -54,7 +57,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "5.0-dev"
+            "dev-master": "5.1-dev"
         }
     }
 }
