@@ -11,7 +11,7 @@
 
 namespace Symfony\Component\Console\Output;
 
-use Symfony\Component\Console\Formatter\OutputFormatter;
+use Symfony\Component\Console\Formatter\NullOutputFormatter;
 use Symfony\Component\Console\Formatter\OutputFormatterInterface;
 
 /**
@@ -24,6 +24,8 @@ use Symfony\Component\Console\Formatter\OutputFormatterInterface;
  */
 class NullOutput implements OutputInterface
 {
+    private $formatter;
+
     /**
      * {@inheritdoc}
      */
@@ -37,8 +39,11 @@ class NullOutput implements OutputInterface
      */
     public function getFormatter()
     {
+        if ($this->formatter) {
+            return $this->formatter;
+        }
         // to comply with the interface we must return a OutputFormatterInterface
-        return new OutputFormatter();
+        return $this->formatter = new NullOutputFormatter();
     }
 
     /**
