@@ -43,7 +43,7 @@ class CountryValidator extends ConstraintValidator
 
         $value = (string) $value;
 
-        if (!Countries::exists($value)) {
+        if ($constraint->alpha3 ? !Countries::alpha3CodeExists($value) : !Countries::exists($value)) {
             $this->context->buildViolation($constraint->message)
                 ->setParameter('{{ value }}', $this->formatValue($value))
                 ->setCode(Country::NO_SUCH_COUNTRY_ERROR)
