@@ -60,6 +60,9 @@ class EmailValidator extends ConstraintValidator
         }
 
         $value = (string) $value;
+        if ('' === $value) {
+            return;
+        }
 
         if (null !== $constraint->normalizer) {
             $value = ($constraint->normalizer)($value);
@@ -70,7 +73,7 @@ class EmailValidator extends ConstraintValidator
         }
 
         if (!\in_array($constraint->mode, Email::$validationModes, true)) {
-            throw new \InvalidArgumentException(sprintf('The "%s::$mode" parameter value is not valid.', \get_class($constraint)));
+            throw new \InvalidArgumentException(sprintf('The "%s::$mode" parameter value is not valid.', get_debug_type($constraint)));
         }
 
         if (Email::VALIDATION_MODE_STRICT === $constraint->mode) {
