@@ -15,6 +15,6 @@ class UnexpectedTypeException extends ValidatorException
 {
     public function __construct($value, string $expectedType)
     {
-        parent::__construct(sprintf('Expected argument of type "%s", "%s" given', $expectedType, \is_object($value) ? \get_class($value) : \gettype($value)));
+        parent::__construct(sprintf('Expected argument of type "%s", "%s" given', $expectedType, get_debug_type($value)));
     }
 }
