@@ -43,7 +43,7 @@ class LinkStub extends ConstStub
 
             return;
         }
-        if (!file_exists($href)) {
+        if (!is_file($href)) {
             return;
         }
         if ($line) {
@@ -72,7 +72,7 @@ class LinkStub extends ConstStub
                 if ('C' === $class[0] && 0 === strpos($class, 'ComposerAutoloaderInit')) {
                     $r = new \ReflectionClass($class);
                     $v = \dirname($r->getFileName(), 2);
-                    if (file_exists($v.'/composer/installed.json')) {
+                    if (is_file($v.'/composer/installed.json')) {
                         self::$vendorRoots[] = $v.\DIRECTORY_SEPARATOR;
                     }
                 }
@@ -91,7 +91,7 @@ class LinkStub extends ConstStub
         }
 
         $parent = $dir;
-        while (!@file_exists($parent.'/composer.json')) {
+        while (!@is_file($parent.'/composer.json')) {
             if (!@file_exists($parent)) {
                 // open_basedir restriction in effect
                 break;
